resolve conflices
