+++ ivy/functional/backends/jax/linear_algebra.py
@@ -101,6 +101,13 @@ def det(x:jnp.array) \
     -> jnp.array:
     return jnp.linalg.det(x)
 
+def cholesky(x: JaxArray, 
+             upper: bool = False) -> JaxArray:
+    if not upper:
+        return jnp.linalg.cholesky(x)
+    else:
+        return jnp.linalg.cholesky(x).T.conj()
+
 
 # Extra #
 # ------#
@@ -122,3 +129,4 @@ def vector_to_skew_symmetric_matrix(vector: JaxArray)\
     row3 = jnp.concatenate((-a2s, a1s, zs), -1)
     # BS x 3 x 3
     return jnp.concatenate((row1, row2, row3), -2)
+
